Renamed the migration db/migrate/20210121070308_create_movies.rb->db/migrate/20210121064446_create_movies.rb, so it gets executed first

+++ db/migrate/20210121052453_create_seats.rb
@@ -4,7 +4,7 @@ class CreateSeats < ActiveRecord::Migration[6.0]
       t.references :auditorium, foreign_key: { to_table: :auditoriums }, index: true
       t.string :row_letter
       t.string :column_number
-      t.integer :base_value
+      t.float :base_value
 
       t.timestamps
     end
